CHATGPT_baseline_suggestion/requirements.txt
elasticsearch>=9.0.0,<10.0.0
Flask>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
